day08: part2
This one was a tough one. Using a naive foldl to calculate the max made
it take effectively forever (extrapolating from some smaller inputs, it
looked like it'd take several years to complete).

Replacing the foldl to get the max at the end with a different
less-deeply-nested-function approach made it actually complete quickly,
but I was not really ready for foldl to be that slow.