Add a faster foldl' implementation
This foldl' implementation is cursed in that it relies on mutating the
current environment, which feels _really bad_, but it's vastly faster.
So much faster I'll take it.

In nickel, creating a new closure is a _really expensive_ operation. It
clones the entire "environment", which includes every closure above
it... which means with foldl' on a 1k element list, you now have a clone
of 1k elements. It's expensive!

This gets around that by constantly mutating and accessing an array
(out) in the current environment, avoiding cloning deeply nested
structures, and then spitting out the answer at the end.

In a proper pure functional language etc etc, this would not work, but
nickel apparently isn't that, and I _need_ all the performance I can
get. Clearly.